feat: add unary message send support + fix signed comparisons
Add support for unary messages (messages with no arguments) in the
Smalltalk implementation and fix a critical bug where comparison
opcodes were using unsigned comparisons instead of signed.

**Unary Message Send:**
- Added send-unary(receiver, selector, cache-id) function
- Works like send-message but without argument parameter
- Uses inline caching for performance (lookup-method-cached)
- Supports message chaining: unary → binary

**New Unary Methods Implemented:**
- negated: arithmetic negation (-x)
- abs: absolute value |x|
- even: test if even (returns 1/0)
- odd: test if odd (returns 1/0)

**Critical Bug Fix - Signed Comparisons:**
- Fixed LT, GT, LTE, GTE opcodes to use int64_t instead of uint64_t
- These were doing unsigned comparisons, breaking negative number logic
- Example: -42 < 0 was returning false (because -42 as unsigned is huge)
- This caused abs(-42) and other negative number operations to fail

**Test Coverage:**
- Test 52.1: Basic unary messages (negated)
- Test 52.2: Additional unary methods (abs)
- Test 52.3: Unary predicates (even, odd)
- Test 52.4: Message chaining (unary + binary)

**Examples:**
```lisp
(send-unary (tag-int 42) negated-sel 20)     ; → -42
(send-unary (tag-int -42) abs-sel 22)        ; → 42
(send-unary (tag-int 42) even-sel 24)        ; → 1 (true)
(send-unary (tag-int 17) odd-sel 27)         ; → 1 (true)

; Chaining: (-10 abs) + 32
(send-message (send-unary (tag-int -10) abs-sel 28)
              sel-plus-id
              (tag-int 32)
              29)                              ; → 42
```

All 180+ tests passing including new unary message tests.

+++ src/stack_vm.hpp
@@ -268,29 +268,29 @@ class StackVM {
                 }
 
                 case Opcode::LT: {
-                    uint64_t b = pop();
-                    uint64_t a = pop();
+                    int64_t b = static_cast<int64_t>(pop());
+                    int64_t a = static_cast<int64_t>(pop());
                     push(a < b ? 1 : 0);
                     break;
                 }
 
                 case Opcode::GT: {
-                    uint64_t b = pop();
-                    uint64_t a = pop();
+                    int64_t b = static_cast<int64_t>(pop());
+                    int64_t a = static_cast<int64_t>(pop());
                     push(a > b ? 1 : 0);
                     break;
                 }
 
                 case Opcode::LTE: {
-                    uint64_t b = pop();
-                    uint64_t a = pop();
+                    int64_t b = static_cast<int64_t>(pop());
+                    int64_t a = static_cast<int64_t>(pop());
                     push(a <= b ? 1 : 0);
                     break;
                 }
 
                 case Opcode::GTE: {
-                    uint64_t b = pop();
-                    uint64_t a = pop();
+                    int64_t b = static_cast<int64_t>(pop());
+                    int64_t a = static_cast<int64_t>(pop());
                     push(a >= b ? 1 : 0);
                     break;
                 }
